Secondary indices have now IndexDefintion covered with tests.
