+++ src/NServiceBus.AzureStoragePersistence.ComponentTests/NServiceBus.AzureStoragePersistence.ComponentTests.csproj
@@ -93,6 +93,7 @@
   <ItemGroup>
     <Compile Include="AzurePersistenceTests.cs" />
     <Compile Include="Persisters\When_using_GlobalConnectionStringConfiguration.cs" />
+    <Compile Include="Sagas\IndexDefinitionTests.cs" />
     <Compile Include="Sagas\When_completing_saga.cs" />
     <Compile Include="Sagas\When_updating_saga.cs" />
     <Compile Include="Sagas\When_saving_saga.cs" />
