+++ src/NServiceBus.AzureStoragePersistence/NServiceBus.AzureStoragePersistence.csproj
@@ -113,6 +113,7 @@
     <Compile Include="SagaPersisters\AzureStorageSagaDefaults.cs" />
     <Compile Include="SagaPersisters\Azure\AzureStorageSagaPersistence.cs" />
     <Compile Include="SagaPersisters\Azure\RetryNeededException.cs" />
+    <Compile Include="SagaPersisters\Azure\SecondaryIndeces\IndexDefintion.cs" />
     <Compile Include="SagaPersisters\Azure\SecondaryIndeces\LRUCache.cs" />
     <Compile Include="SagaPersisters\Azure\SecondaryIndeces\PartitionRowKeyTuple.cs" />
     <Compile Include="SagaPersisters\Azure\SecondaryIndeces\SagaDataSerializer.cs" />
